+++ award_quality_decay_processors.rb
